requirements.txt
gym==0.17.2
Pillow==7.2.0
tf-agents==0.5.0
